some changes to portals
