+++ apps/client/app/(dashboard)/portal/create/form.tsx
@@ -129,7 +129,7 @@ export default function PortalForm() {
       <ActionIcon variant="filled" color="blue" size="lg" onClick={addAddress}>
         <IconPlus />
       </ActionIcon>
-      <div className='my-2'>
+      <div className="my-2">
         <Checkbox
           checked={haveFundingGoal}
           onChange={(event) => {
@@ -152,7 +152,7 @@ export default function PortalForm() {
           </div>
         ) : null}
       </div>
-      <div className='my-2'>
+      <div className="my-2">
         <Checkbox
           checked={haveDeadline}
           onChange={(event) => {
@@ -169,7 +169,7 @@ export default function PortalForm() {
           />
         ) : null}
       </div>
-      {(haveFundingGoal && haveDeadline) ? (
+      {haveFundingGoal && haveDeadline ? (
         <Checkbox
           my="md"
           checked={allowFundsAboveGoal}
