+++ apps/client/app/(dashboard)/prize/create/page.tsx
@@ -1,9 +1,5 @@
 /* eslint-disable @typescript-eslint/no-unused-vars -- I will use them later */
 'use client';
-import ImageComponent from '@/components/Prize/dropzone';
-import usePrizeProposal from '@/components/Prize/hooks/usePrizeProposal';
-import { TextEditor } from '@/components/richtexteditor/textEditor';
-import useAppUser from '@/context/hooks/useAppUser';
 import { Button, Card, Checkbox, NumberInput, SimpleGrid, TextInput, Title } from '@mantine/core';
 import type { FileWithPath } from '@mantine/dropzone';
 import { usePrivy } from '@privy-io/react-auth';
@@ -12,6 +8,10 @@ import router from 'next/router';
 import { useState } from 'react';
 import { toast } from 'sonner';
 import { useMutation } from 'wagmi';
+import useAppUser from '@/context/hooks/useAppUser';
+import { TextEditor } from '@/components/richtexteditor/textEditor';
+import usePrizeProposal from '@/components/Prize/hooks/usePrizeProposal';
+import ImageComponent from '@/components/Prize/dropzone';
 
 function Prize() {
   const [address, setAddress] = useState(['']);
